Refactor: 메인 페이지 컴포넌트 폴더 구조 수정
