--- src/app/(main)/_components/MainSection/Section2/index.tsx
@@ -1,7 +1,0 @@
-export const Section2 = () => {
-  return (
-    <div className="h-[calc(100vh-4rem)] w-full bg-red-100">
-      <div className="">test</div>
-    </div>
-  )
-}
