+++ src/app/(main)/_components/footer/Footer.tsx
