+++ src/app/(main)/_components/footer/FooterDetail.tsx
