+++ src/app/(main)/_components/header/navigation/index.ts
@@ -1,0 +1,2 @@
+export { Navigation } from './Navigation'
+export { MobileNavigation } from './MobileNavigation'
