+++ src/app/(main)/_components/header/profile-button/dropdown-menu/index.ts
@@ -1,0 +1,1 @@
+export { ProfileDropdownMenu } from './DropdownMenu'
