+++ src/app/(main)/_components/header/profile-button/index.ts
@@ -1,0 +1,1 @@
+export { ProfileButton } from './ProfileButton'
