+++ src/app/(main)/_components/index.ts
@@ -1,0 +1,3 @@
+export * from './footer'
+export * from './header'
+export * from './main-section'
