+++ src/app/(main)/_components/main-section/Section1.tsx
