+++ src/app/(main)/_components/main-section/index.ts
@@ -1,0 +1,1 @@
+export { Section1 } from './Section1'
