+++ src/app/(main)/layout.tsx
@@ -1,7 +1,7 @@
 import { Content } from '@/components/common'
 
-import { Footer } from './_components/Footer'
-import { Header } from './_components/Header'
+import { Footer } from './_components'
+import { Header } from './_components/header'
 
 const MainLayout = ({ children }: { children: React.ReactNode }) => {
   return (
